pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "search-query-dsl"
version = "0.1.0"
description = "A Domain-Specific Language for expressing complex database queries as JSON"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "Myron Gourlis" }
]
keywords = ["query", "dsl", "search", "filter", "sqlalchemy", "json"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Database",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Typing :: Typed",
]

# Zero core dependencies
dependencies = []

[project.optional-dependencies]
sqlalchemy = ["sqlalchemy>=2.0"]
geoalchemy = [
    "sqlalchemy>=2.0",
    "geoalchemy2>=0.14",
    "shapely>=2.0",
    "pyproj>=3.0",
]
fastapi = ["pydantic>=2.0"]
django = ["djangorestframework>=3.14"]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
]
all = [
    "search-query-dsl[sqlalchemy,geoalchemy,fastapi,dev]",
]

[project.urls]
Homepage = "https://github.com/enomix-gr/search-query-dsl"
Documentation = "https://github.com/enomix-gr/search-query-dsl#readme"

[tool.hatch.build.targets.wheel]
packages = ["src/search_query_dsl"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
